Account for env vars on running aderyn watch
